+++ sakura_core/outline/FuncInfoArr.cpp
@@ -81,10 +81,10 @@ void FuncInfoArr::AppendData(FuncInfo* pFuncInfo)
 	@date 2002.04.01 YAZAKI 深さ導入
 */
 void FuncInfoArr::AppendData(
-	int				nFuncLineCRLF,		// 関数のある行(CRLF単位)
-	int				nFuncColCRLF,		// 関数のある桁(CRLF単位)
-	int				nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
-	int				nFuncColLAYOUT,		// 関数のある桁(折り返し単位)
+	size_t			nFuncLineCRLF,		// 関数のある行(CRLF単位)
+	size_t			nFuncColCRLF,		// 関数のある桁(CRLF単位)
+	size_t			nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
+	size_t			nFuncColLAYOUT,		// 関数のある桁(折り返し単位)
 	const TCHAR*	pszFuncName,		// 関数名
 	const TCHAR*	pszFileName,		// ファイル名
 	int				nInfo,				// 付加情報
@@ -104,10 +104,10 @@ void FuncInfoArr::AppendData(
 }
 
 void FuncInfoArr::AppendData(
-	int					nFuncLineCRLF,		// 関数のある行(CRLF単位)
-	int					nFuncColCRLF,		// 関数のある桁(CRLF単位)
-	int					nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
-	int					nFuncColLAYOUT,		// 関数のある桁(折り返し単位)
+	size_t				nFuncLineCRLF,		// 関数のある行(CRLF単位)
+	size_t				nFuncColCRLF,		// 関数のある桁(CRLF単位)
+	size_t				nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
+	size_t				nFuncColLAYOUT,		// 関数のある桁(折り返し単位)
 	const NOT_TCHAR*	pszFuncName,		// 関数名
 	const NOT_TCHAR*	pszFileName,		// ファイル名
 	int					nInfo,				// 付加情報
@@ -126,8 +126,8 @@ void FuncInfoArr::AppendData(
 
 
 void FuncInfoArr::AppendData(
-	int				nFuncLineCRLF,		// 関数のある行(CRLF単位)
-	int				nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
+	size_t			nFuncLineCRLF,		// 関数のある行(CRLF単位)
+	size_t			nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
 	const TCHAR*	pszFuncName,		// 関数名
 	int				nInfo,				// 付加情報
 	size_t			nDepth				// 深さ
@@ -145,8 +145,8 @@ void FuncInfoArr::AppendData(
 }
 
 void FuncInfoArr::AppendData(
-	int					nFuncLineCRLF,		// 関数のある行(CRLF単位)
-	int					nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
+	size_t				nFuncLineCRLF,		// 関数のある行(CRLF単位)
+	size_t				nFuncLineLAYOUT,	// 関数のある行(折り返し単位)
 	const NOT_TCHAR*	pszFuncName,		// 関数名
 	int					nInfo,				// 付加情報
 	size_t				nDepth				// 深さ
