+++ sakura_core/outline/FuncInfoArr.h
@@ -39,10 +39,10 @@ public:
 	~FuncInfoArr();	// FuncInfoArrクラス消滅
 	FuncInfo* GetAt(size_t);		// 0<=の指定番号のデータを返す
 	void AppendData(FuncInfo*);	// 配列の最後にデータを追加する
-	void AppendData(int, int, const TCHAR*, int, size_t nDepth = 0);		// 配列の最後にデータを追加する 2002.04.01 YAZAKI 深さ導入
-	void AppendData(int, int, const NOT_TCHAR*, int, size_t nDepth = 0);	// 配列の最後にデータを追加する 2002.04.01 YAZAKI 深さ導入
-	void AppendData(int nLogicLine, int nLogicCol, int nLayoutLine, int nLayoutCol, const TCHAR*, const TCHAR*, int, size_t nDepth = 0);	/* 配列の最後にデータを追加する 2010.03.01 syat 桁導入*/
-	void AppendData(int nLogicLine, int nLogicCol, int nLayoutLine, int nLayoutCol, const NOT_TCHAR*, const NOT_TCHAR*, int, size_t nDepth = 0);	/* 配列の最後にデータを追加する 2010.03.01 syat 桁導入*/
+	void AppendData(size_t, size_t, const TCHAR*, int, size_t nDepth = 0);		// 配列の最後にデータを追加する 2002.04.01 YAZAKI 深さ導入
+	void AppendData(size_t, size_t, const NOT_TCHAR*, int, size_t nDepth = 0);	// 配列の最後にデータを追加する 2002.04.01 YAZAKI 深さ導入
+	void AppendData(size_t nLogicLine, size_t nLogicCol, size_t nLayoutLine, size_t nLayoutCol, const TCHAR*, const TCHAR*, int, size_t nDepth = 0);	/* 配列の最後にデータを追加する 2010.03.01 syat 桁導入*/
+	void AppendData(size_t nLogicLine, size_t nLogicCol, size_t nLayoutLine, size_t nLayoutCol, const NOT_TCHAR*, const NOT_TCHAR*, int, size_t nDepth = 0);	/* 配列の最後にデータを追加する 2010.03.01 syat 桁導入*/
 	size_t	GetNum(void) {	return nFuncInfoArrNum; }	// 配列要素数を返す
 	void Empty(void);
 	void DUMP(void);
