+++ sakura_core/recent/RecentCurDir.cpp
@@ -53,7 +53,7 @@ RecentCurDir::RecentCurDir()
 
 	@note	取得後のポインタはユーザ管理の構造体にキャストして参照してください。
 */
-const TCHAR* RecentCurDir::GetItemText(int nIndex) const
+const TCHAR* RecentCurDir::GetItemText(size_t nIndex) const
 {
 	return *GetItem(nIndex);
 }
