+++ sakura_core/recent/RecentFolder.cpp
@@ -52,7 +52,7 @@ RecentFolder::RecentFolder()
 
 	@note	取得後のポインタはユーザ管理の構造体にキャストして参照してください。
 */
-const TCHAR* RecentFolder::GetItemText(int nIndex) const
+const TCHAR* RecentFolder::GetItemText(size_t nIndex) const
 {
 	return *GetItem(nIndex);
 }
