+++ sakura_core/recent/RecentGrepFolder.cpp
@@ -52,7 +52,7 @@ RecentGrepFolder::RecentGrepFolder()
 
 	@note	取得後のポインタはユーザ管理の構造体にキャストして参照してください。
 */
-const TCHAR* RecentGrepFolder::GetItemText(int nIndex) const
+const TCHAR* RecentGrepFolder::GetItemText(size_t nIndex) const
 {
 	return *GetItem(nIndex);
 }
